Implement Firestore integration for product and wishlist management; update UI components for better user experience

+++ app/src/main/java/com/jason/supermarketapp/ui/viewmodels/ProductDetailsViewModel.kt
@@ -2,10 +2,15 @@ package com.jason.supermarketapp.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.jason.supermarketapp.R
 import com.jason.supermarketapp.data.repositories.WishlistRepository
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 
 class ProductDetailsViewModel : ViewModel() {
@@ -14,26 +19,28 @@ class ProductDetailsViewModel : ViewModel() {
 
     private val _isInWishlist = MutableStateFlow(false)
     val isInWishlist: StateFlow<Boolean> = _isInWishlist.asStateFlow()
+    private val _uiMessage = MutableSharedFlow<Int>()
+    val uiMessage: SharedFlow<Int> = _uiMessage.asSharedFlow()
 
     fun checkWishlistStatus(productId: String) {
         viewModelScope.launch {
-            wishlistRepository.getWishlistItem(productId).collect { item ->
-                _isInWishlist.value = item != null
+            wishlistRepository.isProductInWishlistFlow(productId)
+                .collectLatest { isProductInWishlist ->
+                    _isInWishlist.value = isProductInWishlist
+                }
             }
-        }
     }
 
-    fun addToWishlist(productId: String) {
+    fun toggleWishlistStatus(productId: String) {
         viewModelScope.launch {
-            wishlistRepository.addWishlistItem(productId)
-            _isInWishlist.value = true
-        }
-    }
-
-    fun removeFromWishlist(productId: String) {
-        viewModelScope.launch {
-            wishlistRepository.removeWishlistItem(productId)
-            _isInWishlist.value = false
+            val isInWishlist = isInWishlist.value
+            if (isInWishlist) {
+                wishlistRepository.removeWishlistItem(productId)
+                _uiMessage.emit(R.string.removed_from_wishlist)
+            } else {
+                wishlistRepository.addWishlistItem(productId)
+                _uiMessage.emit(R.string.added_to_wishlist)
+            }
         }
     }
 }
